Clean up diff/index.js to ensure nothing uses it
