+++ src/diff/unmount.js
@@ -1,0 +1,49 @@
+import options from '../options';
+import { removeNode } from '../util';
+import { applyRef } from './refs';
+
+/**
+ * Unmount a virtual node from the tree and apply DOM changes
+ * @param {import('../internal').VNode} vnode The virtual node to unmount
+ * @param {import('../internal').VNode} parentVNode The parent of the VNode that
+ * initiated the unmount
+ * @param {boolean} [skipRemove] Flag that indicates that a parent node of the
+ * current element is already detached from the DOM.
+ */
+export function unmount(vnode, parentVNode, skipRemove) {
+	let r;
+	if (options.unmount) options.unmount(vnode);
+
+	if ((r = vnode.ref)) {
+		if (!r.current || r.current === vnode._dom) applyRef(r, null, parentVNode);
+	}
+
+	let dom;
+	if (!skipRemove && typeof vnode.type != 'function') {
+		skipRemove = (dom = vnode._dom) != null;
+	}
+
+	// Must be set to `undefined` to properly clean up `_nextDom`
+	// for which `null` is a valid value. See comment in `create-element.js`
+	vnode._dom = vnode._nextDom = undefined;
+
+	if ((r = vnode._component) != null) {
+		if (r.componentWillUnmount) {
+			try {
+				r.componentWillUnmount();
+			} catch (e) {
+				options._catchError(e, parentVNode);
+			}
+		}
+
+		r.base = r._parentDom = null;
+	}
+
+	if ((r = vnode._children)) {
+		for (let i = 0; i < r.length; i++) {
+			if (r[i]) unmount(r[i], parentVNode, skipRemove);
+		}
+	}
+
+	if (dom != null) removeNode(dom);
+}
